% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIYprofile.R
\name{DIYprofile}
\alias{DIYprofile}
\title{Extract counts per million (CPM) from a DGEList object, convert to a tibble, pivot_longer to make tidy, and then plot with ggplot to see distribution of CPM}
\usage{
DIYprofile(data, samples)
}
\arguments{
\item{data}{A object of class DGEList.}

\item{samples}{A character vector of sample names.}
}
\value{
A violin plot of CPM (y axis) for each sample (x axis)
}
\description{
Extract counts per million (CPM) from a DGEList object, convert to a tibble, pivot_longer to make tidy, and then plot with ggplot to see distribution of CPM
}
