% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myDGEList}
\alias{myDGEList}
\title{Gene expression data for human cutaneous leishmaniasis}
\format{
A data frame with 35643 rows (genes) and 10 variables (samples):
\describe{
\item{HS..}{healthy subject}
\item{CL..}{cutaneous leishmaniasis}
...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA525604}
}
\usage{
myDGEList
}
\description{
A DGEList object containing counts of reads mapping (via Kallisto) to human genes
}
\keyword{datasets}
