% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targets}
\alias{targets}
\title{Sample metadata for each patient sample}
\format{
A tibble containing 10 observations of 3 variables:
\describe{
\item{sample}{sample ID}
\item{sra_accession}{accession number for the sequence read archive}
\item{group}{healthy or disease (cutaneous leishmaniasis)}
...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA525604}
}
\usage{
targets
}
\description{
Sample metadata for each patient sample
}
\keyword{datasets}
